+++ src/main/java/grit/stockIt/domain/ranking/service/RankingService.java
@@ -8,6 +8,7 @@ import grit.stockIt.domain.contest.entity.Contest;
 import grit.stockIt.domain.contest.repository.ContestRepository;
 import grit.stockIt.domain.matching.repository.RedisMarketDataRepository;
 import grit.stockIt.domain.mission.service.MissionService;
+import grit.stockIt.domain.mission.service.MissionQueryService;
 import grit.stockIt.domain.mission.dto.UserTierStatusResponse;
 import grit.stockIt.domain.ranking.dto.MyRankResponse;
 import grit.stockIt.domain.ranking.dto.RankingItemResponse;
@@ -46,6 +47,7 @@ public class RankingService {
     private final AccountStockRepository accountStockRepository;
     private final ContestRepository contestRepository;
     private final MissionService missionService;
+    private final MissionQueryService missionQueryService;
     private final RedisMarketDataRepository redisMarketDataRepository;
     private final StockDetailService stockDetailService;
     private final Environment environment;
@@ -278,7 +280,7 @@ public class RankingService {
      */
     private String getTierForMember(grit.stockIt.domain.member.entity.Member member) {
         try {
-            UserTierStatusResponse tierInfo = missionService.getTierInfo(member.getEmail());
+            UserTierStatusResponse tierInfo = missionQueryService.getTierInfo(member.getEmail());
             return tierInfo.getCurrentTier();
         } catch (Exception e) {
             log.warn("티어 조회 실패 (memberId: {}): {}", member.getMemberId(), e.getMessage());
